+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Unreleased
 
+### Changed
+- When saving a Neo field's settings, existing block type groups are now properly resaved, not just deleted and recreated
+
 ### Fixed
+- Fixed an error, involving block type group resaving, that could occur when running `./craft project-config/apply` when Craft's `allowAdminChanges` setting is disabled
 - Fixed an error that occurred with Craft 3.5 / Neo 2.8 when rebuilding the project config, if any Neo block type field layout consisted only of blank tab(s)
 - Fixed a bug with previous Neo 2.8 releases where default values on dropdown fields were not being applied to new Neo blocks
 - Fixed a potential GraphQL "Failed to validate the GQL Schema" error when querying for Neo field data